refactor: solved issues with project structure and dependencies
