+++ src/Application/Services/ServiceRead.cs
@@ -10,7 +10,9 @@ using System.Threading.Tasks;
 using IAVH.BioTablero.CM.Application.Interfaces.General;
 using IAVH.BioTablero.CM.Application.Specifications;
 using IAVH.BioTablero.CM.Core.Helpers.General;
-using IAVH.BioTablero.CM.Core.interfaces;
+using IAVH.BioTablero.CM.Core.Interfaces.DTOs;
+using IAVH.BioTablero.CM.Core.Interfaces.Entities;
+using IAVH.BioTablero.CM.Core.Interfaces.Repositories;
 
 using Microsoft.AspNetCore.OData.Query;
 using Microsoft.EntityFrameworkCore;
